Remove stuff from home controller
