+++ DapperDoodle.TestProject/Models/Person.cs
@@ -1,0 +1,9 @@
+namespace TestProject.Models
+{
+    public class Person
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public string Surname { get; set; }
+    }
+}
